requirements.txt
natsort==8.1.0
Pillow==9.0.1
